+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Add manual GitHub Actions tests for pysatMadrigal RC
   * Update GitHub Actions workflows for newer versions of pip
   * Added clean warning test attributes to all Instruments
+  * Updated documentation to comply with current Ecosystem guidelines
+  * Replaced setup.py with pyproject.toml
 
 [0.1.0] - 2023-04-11
 --------------------