+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-[0.2.0] - 2024-01-10
+[0.2.0] - 2024-XX-XX
 --------------------
 * Enhancements
   * Moved the OMNI-2 IMF, Dst, and Geomagnetic index Instruments from the